Color output of Memoizing context
