+++ flo-freezer/src/main/java/io/rouz/flo/freezer/LoggingListener.java
@@ -1,12 +1,9 @@
 package io.rouz.flo.freezer;
 
-import static org.fusesource.jansi.Ansi.Color.CYAN;
-import static org.fusesource.jansi.Ansi.Color.WHITE;
-import static org.fusesource.jansi.Ansi.ansi;
+import static io.rouz.flo.Util.colored;
 
 import io.rouz.flo.TaskId;
 import io.rouz.flo.context.InstrumentedContext;
-import org.fusesource.jansi.Ansi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -25,19 +22,5 @@ public class LoggingListener implements InstrumentedContext.Listener {
   @Override
   public void status(TaskId task, Phase phase) {
     LOG.info("{} :: {}", colored(task), phase);
-  }
-
-  public static Ansi colored(TaskId taskId) {
-    final String id = taskId.toString();
-    final int openParen = id.indexOf('(');
-    final int closeParen = id.lastIndexOf(')');
-    final int hashPos = id.lastIndexOf('#');
-
-    return ansi()
-        .fg(CYAN).a(id.substring(0, openParen + 1))
-        .reset().a(id.substring(openParen + 1, closeParen))
-        .fg(CYAN).a(id.substring(closeParen, hashPos))
-        .fg(WHITE).a(id.substring(hashPos))
-        .reset();
   }
 }
